GroupCW/AzureFunctions/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
azure-cognitiveservices-speech==1.34.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.29.5
azure-cosmos==4.5.1
azure-functions==1.17.0
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.19.0
bcrypt==3.2.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.7
decorator==4.4.2
distlib==0.3.7
distro==1.9.0
filelock==3.12.4
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
isodate==0.6.1
moviepy==1.0.3
msrest==0.7.1
numpy==1.26.2
oauthlib==3.2.2
openai==1.6.1
Pillow==10.1.0
platformdirs==3.11.0
proglog==0.1.10
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
tinydb==4.8.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.6
virtualenv==20.24.5
